chore: 기본 역량 키워드 세팅 임시 API 구현(#144)
