+++ Api-Module/src/main/kotlin/com/bamyanggang/apimodule/domain/experience/presentation/ExperienceController.kt
@@ -8,7 +8,6 @@ import com.bamyanggang.apimodule.domain.experience.application.service.Experienc
 import com.bamyanggang.apimodule.domain.experience.application.service.ExperienceDeleteService
 import com.bamyanggang.apimodule.domain.experience.application.service.ExperienceEditService
 import com.bamyanggang.apimodule.domain.experience.application.service.ExperienceGetService
-import com.bamyanggang.domainmodule.domain.strongpoint.aggregate.Keyword
 import com.bamyanggang.persistence.strongpoint.jpa.repository.KeywordJpaRepository
 import com.bamyanggang.persistence.strongpoint.mapper.KeywordMapper
 import org.springframework.web.bind.annotation.*
@@ -20,24 +19,7 @@ class ExperienceController(
     private val experienceDeleteService: ExperienceDeleteService,
     private val experienceEditService: ExperienceEditService,
     private val experienceGetService: ExperienceGetService,
-    private val keywordJpaRepository: KeywordJpaRepository,
-    private val keywordMapper: KeywordMapper
 ) {
-    data class keyword(
-        val id: UUID,
-        val name: String
-    )
-
-    @PostMapping("/api/set-data")
-    fun setData(words : List<keyword>) {
-        val datas = words.map { Keyword(it.id, it.name) }.map { keywordMapper.toJpaEntity(it) }
-        keywordJpaRepository.saveAll(datas)
-    }
-
-    @GetMapping("/api/default-points")
-    fun getAllPoints() : List<Keyword> {
-        return keywordJpaRepository.findAll().map { keywordMapper.toDomainEntity(it) }
-    }
 
     @GetMapping(ExperienceApi.BOOKMARK_EXPERIENCE_URL)
     fun getBookMarkExperiences(
