+++ Api-Module/src/main/kotlin/com/bamyanggang/apimodule/domain/strongpoint/presentation/StrongPointController.kt
@@ -5,14 +5,20 @@ import com.bamyanggang.apimodule.domain.strongpoint.application.dto.GetStrongPoi
 import com.bamyanggang.apimodule.domain.strongpoint.application.service.StrongPointCreateService
 import com.bamyanggang.apimodule.domain.strongpoint.application.service.StrongPointDeleteService
 import com.bamyanggang.apimodule.domain.strongpoint.application.service.StrongPointGetService
+import com.bamyanggang.domainmodule.domain.strongpoint.aggregate.Keyword
+import com.bamyanggang.persistence.strongpoint.jpa.repository.KeywordJpaRepository
+import com.bamyanggang.persistence.strongpoint.mapper.KeywordMapper
 import org.springframework.web.bind.annotation.*
 import java.util.*
+
 
 @RestController
 class StrongPointController(
     private val strongPointCreateService: StrongPointCreateService,
     private val strongPointDeleteService: StrongPointDeleteService,
-    private val strongPointGetService: StrongPointGetService
+    private val strongPointGetService: StrongPointGetService,
+    private val keywordJpaRepository: KeywordJpaRepository,
+    private val keywordMapper: KeywordMapper
 ) {
     @GetMapping(StrongPointApi.BASE_URL)
     fun getAllStrongPoints(): GetStrongPoint.Response {
@@ -27,5 +33,29 @@ class StrongPointController(
     @DeleteMapping(StrongPointApi.STRONG_POINT_PATH_VARIABLE_URL)
     fun deleteStrongPoint(@PathVariable("strongPointId") strongPointId: UUID) {
         strongPointDeleteService.deleteStrongPoint(strongPointId)
+    }
+
+    // 기본 역량 키워드 데이터 세팅 임시 API
+    data class Data(
+        val points : List<KeywordData>
+    ) {
+
+        data class KeywordData (
+            val id: UUID,
+            val name: String
+        )
+    }
+
+    @PostMapping("/api/set-data")
+    fun setData(@RequestBody data : Data) {
+        val datas = data.points.map { Keyword(it.id, it.name) }
+        val jpaDatas = datas.map { keywordMapper.toJpaEntity(it) }
+
+        keywordJpaRepository.saveAll(jpaDatas)
+    }
+
+    @GetMapping("/api/default-points")
+    fun getAllPoints() : List<Keyword> {
+        return keywordJpaRepository.findAll().map { keywordMapper.toDomainEntity(it) }
     }
 }
